+++ test/e2e/utils/anpspecbuilder.go
@@ -95,7 +95,7 @@ func (b *AntreaNetworkPolicySpecBuilder) GetAppliedToPeer(podSelector map[string
 }
 
 func (b *AntreaNetworkPolicySpecBuilder) AddIngress(protoc v1.Protocol,
-	port *int, portName *string, endPort *int32, cidr *string,
+	port *int32, portName *string, endPort *int32, cidr *string,
 	podSelector map[string]string, nsSelector map[string]string,
 	podSelectorMatchExp []metav1.LabelSelectorRequirement, nsSelectorMatchExp []metav1.LabelSelectorRequirement,
 	ruleAppliedToSpecs []ANPAppliedToSpec, action secv1alpha1.RuleAction, name string) *AntreaNetworkPolicySpecBuilder {
@@ -165,7 +165,7 @@ func (b *AntreaNetworkPolicySpecBuilder) AddIngress(protoc v1.Protocol,
 	if port != nil || endPort != nil {
 		var pVal *intstr.IntOrString
 		if port != nil {
-			pVal = &intstr.IntOrString{IntVal: int32(*port)}
+			pVal = &intstr.IntOrString{IntVal: *port}
 		}
 		ports = []secv1alpha1.NetworkPolicyPort{
 			{
@@ -188,7 +188,7 @@ func (b *AntreaNetworkPolicySpecBuilder) AddIngress(protoc v1.Protocol,
 }
 
 func (b *AntreaNetworkPolicySpecBuilder) AddEgress(protoc v1.Protocol,
-	port *int, portName *string, endPort *int32, cidr *string,
+	port *int32, portName *string, endPort *int32, cidr *string,
 	podSelector map[string]string, nsSelector map[string]string,
 	podSelectorMatchExp []metav1.LabelSelectorRequirement, nsSelectorMatchExp []metav1.LabelSelectorRequirement,
 	ruleAppliedToSpecs []ANPAppliedToSpec, action secv1alpha1.RuleAction, name string) *AntreaNetworkPolicySpecBuilder {
